requirements.txt
Django==3.2.13
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
pyotp==2.6.0
djangorestframework-api-key==2.1.0
celery== 5.2.6
django-celery-beat==2.2.1
django_redis==5.0.0
# for celery importy error install
django-extensions==2.2.9
importlib-metadata==4.8.3